commit 3

+++ client/src/components/dialog/Delete.jsx
@@ -1,5 +1,4 @@
 import React, { Component } from 'react'
-
 import {
     Dialog,
     DialogActions,
@@ -7,40 +6,47 @@ import {
     Button,
     DialogContentText,
 } from '@mui/material'
-
 export default class Delete extends Component {
-    state = {
-        openDelete: false,
-        id: '',
-    }
-
-    componentDidUpdate() {
-        if (
-            this.props.openDelete !== this.state.openDelete ||
-            this.props.id !== this.state.id
-        ) {
-            this.setState({
-                openDelete: this.props.openDelete,
-                id: this.props.id,
-            })
-        }
-    }
-
     render() {
         return (
             <div>
                 <Dialog
-                    open={this.state.openDelete}
+                    open={this.props.openDelete || this.props.openLock}
                     onClose={this.props.handleCloseDelete}
+                    fullWidth={true}
                 >
-                    <DialogTitle color={"error"}>
-                        Xác nhận xoá tài khoản {this.props.username}
+                    <DialogTitle
+                        color={
+                            this.props.openLock &&
+                            this.props.userStatus === 'Không hoạt động'
+                                ? 'success'
+                                : 'error'
+                        }
+                    >
+                        Xác nhận{' '}
+                        {this.props.openDelete
+                            ? 'xoá'
+                            : this.props.openLock &&
+                              this.props.userStatus === 'Không hoạt động'
+                            ? 'kích hoạt'
+                            : 'khoá'}{' '}
+                        tài khoản {this.props.username}
                     </DialogTitle>
                     <DialogContentText align="center">
                         <b>
-                            Bạn có muốn chắc chắn xoá tài khoản này, hành động
-                            này sẽ khiến người dùng không thể truy cập vào hệ
-                            thống!
+                            Bạn có muốn chắc chắn{' '}
+                            {this.props.openDelete
+                                ? 'xoá'
+                                : this.props.openLock &&
+                                  this.props.userStatus === 'Không hoạt động'
+                                ? 'kích hoạt'
+                                : 'khoá'}{' '}
+                            tài khoản này, hành động này sẽ khiến người dùng{' '}
+                            {this.props.openLock &&
+                            this.props.userStatus === 'Không hoạt động'
+                                ? 'có'
+                                : 'không'}{' '}
+                            thể truy cập vào hệ thống!
                         </b>
                     </DialogContentText>
                     <DialogActions>
@@ -49,11 +55,40 @@ export default class Delete extends Component {
                         </Button>
                         <Button
                             variant="contained"
-                            color="error"
+                            color={
+                                this.props.openLock &&
+                                this.props.userStatus === 'Không hoạt động'
+                                    ? 'success'
+                                    : 'error'
+                            }
                             onClick={() => {
-                                this.props.deleteUser({
-                                    id: this.state.id,
-                                })
+                                if (this.props.openDelete) {
+                                    this.props.deleteUser({
+                                        id: this.props.id,
+                                    })
+                                } else if (
+                                    this.props.openLock &&
+                                    this.props.userStatus === 'Không hoạt động'
+                                ) {
+                                    this.props.updateUser({
+                                        id: this.props.id,
+                                        username: this.props.username,
+                                        password: this.props.password,
+                                        dateOfBirth: this.props.dateOfBirth,
+                                        userStatus: 'Hoạt động',
+                                    })
+                                } else if (
+                                    this.props.openLock &&
+                                    this.props.userStatus === 'Hoạt động'
+                                ) {
+                                    this.props.updateUser({
+                                        id: this.props.id,
+                                        username: this.props.username,
+                                        password: this.props.password,
+                                        dateOfBirth: this.props.dateOfBirth,
+                                        userStatus: 'Không hoạt động',
+                                    })
+                                }
                                 this.props.handleCloseDelete()
                             }}
                         >
